refactor: fold tier service into server keys
